micro: extract pi runtime helper seams
