+++ frontend/src/lib/agent/pi-runtime.ts
@@ -1,7 +1,5 @@
 import { EventEmitter } from "node:events";
-import { existsSync, readFileSync } from "node:fs";
-import { chmod, mkdir, realpath, stat, writeFile } from "node:fs/promises";
-import { homedir } from "node:os";
+import { chmod, mkdir, writeFile } from "node:fs/promises";
 import path from "node:path";
 import { spawn, type ChildProcessWithoutNullStreams } from "node:child_process";
 import { getApiSettings, type ApiSettings } from "@/lib/api-settings";
@@ -9,7 +7,23 @@ import { resolveDataDir } from "@/lib/data-dir";
 import { normalizeOpenAIModels, modelsToPiModels, type AgentModel } from "./models";
 import { isAgentEndEvent } from "./pi-events";
 import { piPathEnv, resolvePiLaunchCommand } from "./pi-binary";
-import { listProjectsFromStore } from "./projects-store";
+import {
+  deriveFrontendBase,
+  normalizeBackendUrl,
+  pluginFingerprint,
+  pluginMcpConfigs,
+  pluginNameMatches,
+  pluginSkillPaths,
+  resolveAgentCwd,
+  resolveBrowserExtensionPath,
+  resolveComputerUseApp,
+  resolveMcpExtensionPath,
+  resolveTimeoutExtensionPath,
+  selectedSkillPaths,
+  uniqueExistingPaths,
+  type RuntimePluginRef,
+  type RuntimeStartOptions,
+} from "./pi-runtime-helpers";
 
 const PROVIDER_ID = "vllm-studio";
 const DEFAULT_SESSION_ID = "default";
@@ -37,169 +51,6 @@ type PendingCommand = {
   reject: (error: Error) => void;
 };
 
-type RuntimePluginRef = {
-  id?: string;
-  name?: string;
-  path?: string;
-  skillPath?: string;
-  mcpConfigPath?: string;
-  appConfigPath?: string;
-  appIds?: string[];
-  appPath?: string;
-};
-
-type RuntimeSkillRef = {
-  id?: string;
-  name?: string;
-  path?: string;
-};
-
-type RuntimeStartOptions = {
-  browserToolEnabled?: boolean;
-  plugins?: RuntimePluginRef[];
-  skills?: RuntimeSkillRef[];
-};
-
-function normalizeBackendUrl(value: string): string {
-  return value.trim().replace(/\/+$/, "");
-}
-
-function resolveDefaultAgentCwd(): string {
-  // Explicit override always wins.
-  if (process.env.VLLM_STUDIO_AGENT_CWD) return process.env.VLLM_STUDIO_AGENT_CWD;
-
-  // In a packaged Electron app, process.cwd() is "/" — useless as a working
-  // directory for the coding agent. If the renderer hasn't picked a project,
-  // fall back to the most recently added project on disk, then to $HOME.
-  try {
-    const projects = listProjectsFromStore();
-    const usable = projects.find((entry) => entry.exists);
-    if (usable) return usable.path;
-  } catch {
-    // ignore — projects.json may not exist yet
-  }
-
-  // Dev: if cwd is the frontend/ dir, use the repo root.
-  const cwd = process.cwd();
-  if (path.basename(cwd) === "frontend") return path.resolve(cwd, "..");
-
-  // Bare process.cwd() === "/" is unusable; prefer $HOME.
-  if (cwd === "/" || cwd === "") return homedir();
-  return cwd;
-}
-
-function expandHome(value: string): string {
-  if (value === "~") return homedir();
-  if (value.startsWith(`~${path.sep}`)) return path.join(homedir(), value.slice(2));
-  return value;
-}
-
-async function resolveAgentCwd(input?: string): Promise<string> {
-  const defaultCwd = resolveDefaultAgentCwd();
-  const raw = input?.trim() || defaultCwd;
-  const expanded = expandHome(raw);
-  const candidate = path.isAbsolute(expanded) ? expanded : path.resolve(defaultCwd, expanded);
-  const resolved = await realpath(candidate);
-  const info = await stat(resolved);
-  if (!info.isDirectory()) {
-    throw new Error(`Agent cwd is not a directory: ${resolved}`);
-  }
-  return resolved;
-}
-
-// Locate bundled Pi extensions. In dev they sit next to the source files;
-// in a packaged Electron app they ship under
-// process.resourcesPath/desktop/resources/pi-extensions/. We accept either.
-function resolveBundledPiExtensionPath(fileName: string, envOverride?: string): string | null {
-  const candidates = [
-    envOverride,
-    process.resourcesPath
-      ? path.join(process.resourcesPath, "desktop", "resources", "pi-extensions", fileName)
-      : null,
-    path.resolve(process.cwd(), "frontend", "desktop", "resources", "pi-extensions", fileName),
-    path.resolve(process.cwd(), "desktop", "resources", "pi-extensions", fileName),
-    path.resolve(
-      process.cwd(),
-      "..",
-      "frontend",
-      "desktop",
-      "resources",
-      "pi-extensions",
-      fileName,
-    ),
-  ].filter((value): value is string => Boolean(value));
-  for (const candidate of candidates) {
-    if (existsSync(candidate)) return candidate;
-  }
-  return null;
-}
-
-function resolveBrowserExtensionPath(): string | null {
-  return resolveBundledPiExtensionPath(
-    "browser.ts",
-    process.env.VLLM_STUDIO_BROWSER_EXTENSION_PATH,
-  );
-}
-
-function resolveTimeoutExtensionPath(): string | null {
-  return resolveBundledPiExtensionPath(
-    "vllm-studio-timeouts.ts",
-    process.env.VLLM_STUDIO_TIMEOUT_EXTENSION_PATH,
-  );
-}
-
-function resolveMcpExtensionPath(): string | null {
-  return resolveBundledPiExtensionPath("mcp-plugin.ts", process.env.VLLM_STUDIO_MCP_EXTENSION_PATH);
-}
-
-function pluginNameMatches(plugin: RuntimePluginRef, needle: string): boolean {
-  return [
-    plugin.id,
-    plugin.name,
-    plugin.path,
-    plugin.skillPath,
-    plugin.mcpConfigPath,
-    plugin.appConfigPath,
-    plugin.appPath,
-  ]
-    .filter((value): value is string => Boolean(value))
-    .some((value) => value.toLowerCase().includes(needle));
-}
-
-function pluginFingerprint(options: RuntimeStartOptions): string {
-  const names = (options.plugins ?? [])
-    .map(
-      (plugin) =>
-        `${plugin.name ?? ""}:${plugin.path ?? ""}:${plugin.skillPath ?? ""}:${plugin.mcpConfigPath ?? ""}:${plugin.appConfigPath ?? ""}:${plugin.appIds?.join(",") ?? ""}:${plugin.appPath ?? ""}`,
-    )
-    .sort();
-  const skills = (options.skills ?? [])
-    .map((skill) => `${skill.name ?? ""}:${skill.path ?? ""}`)
-    .sort();
-  return JSON.stringify({
-    browser: options.browserToolEnabled === true,
-    plugins: names,
-    skills,
-  });
-}
-
-function resolveComputerUseApp(plugins: RuntimePluginRef[]): string | null {
-  const selected = plugins.find((plugin) => pluginNameMatches(plugin, "computer-use"));
-  const candidates = [
-    selected?.appPath,
-    selected?.path && !selected.path.endsWith(".app")
-      ? path.join(selected.path, "Codex Computer Use.app")
-      : null,
-    selected?.path,
-    "/Applications/Codex.app/Contents/Resources/plugins/openai-bundled/plugins/computer-use/Codex Computer Use.app",
-    path.join(resolveDataDir(), "computer-use", "Codex Computer Use.app"),
-    path.join(homedir(), ".codex", "computer-use", "Codex Computer Use.app"),
-  ].filter((value): value is string => Boolean(value));
-  return (
-    candidates.find((candidate) => candidate.endsWith(".app") && existsSync(candidate)) ?? null
-  );
-}
-
 function launchComputerUseApp(plugins: RuntimePluginRef[]) {
   if (process.platform !== "darwin") return;
   const appPath = resolveComputerUseApp(plugins);
@@ -209,96 +60,6 @@ function launchComputerUseApp(plugins: RuntimePluginRef[]) {
     stdio: "ignore",
   });
   child.unref();
-}
-
-function pluginSkillPaths(plugins: RuntimePluginRef[]): string[] {
-  return uniqueExistingPaths(
-    plugins.flatMap((plugin) => [
-      plugin.skillPath,
-      plugin.path && !plugin.path.endsWith(".app") ? path.join(plugin.path, "skills") : null,
-    ]),
-  );
-}
-
-function selectedSkillPaths(skills: RuntimeSkillRef[]): string[] {
-  return uniqueExistingPaths(skills.map((skill) => skill.path));
-}
-
-function uniqueExistingPaths(values: Array<string | null | undefined>): string[] {
-  const seen = new Set<string>();
-  return values.filter((value): value is string => {
-    if (!value || !existsSync(value)) return false;
-    const resolved = path.resolve(value);
-    if (seen.has(resolved)) return false;
-    seen.add(resolved);
-    return true;
-  });
-}
-
-function isLaunchConstrainedComputerUseMcp(configPath: string): boolean {
-  try {
-    const parsed = JSON.parse(readFileSync(configPath, "utf8")) as {
-      mcpServers?: Record<string, { command?: unknown; args?: unknown }>;
-    };
-    return Object.entries(parsed.mcpServers ?? {}).some(([name, server]) => {
-      const marker =
-        `${name} ${String(server.command ?? "")} ${Array.isArray(server.args) ? server.args.join(" ") : ""}`.toLowerCase();
-      return marker.includes("computer-use") || marker.includes("skycomputeruseclient");
-    });
-  } catch {
-    return false;
-  }
-}
-
-function shouldLoadMcpConfig(plugin: RuntimePluginRef, configPath: string): boolean {
-  if (process.env.VLLM_STUDIO_ENABLE_CODEX_COMPUTER_USE_MCP === "1") return true;
-  if (isLocalComputerUseHelper(plugin, configPath)) return true;
-  return !(
-    pluginNameMatches(plugin, "computer-use") && isLaunchConstrainedComputerUseMcp(configPath)
-  );
-}
-
-function isLocalComputerUseHelper(plugin: RuntimePluginRef, configPath: string): boolean {
-  return (
-    pluginNameMatches(plugin, "computer-use") &&
-    localComputerUseRoots().some((root) => isPathInside(configPath, root))
-  );
-}
-
-function localComputerUseRoots(): string[] {
-  return [
-    path.join(resolveDataDir(), "computer-use"),
-    path.join(homedir(), ".codex", "computer-use"),
-  ];
-}
-
-function isPathInside(candidate: string, root: string): boolean {
-  const relative = path.relative(path.resolve(root), path.resolve(candidate));
-  return relative === "" || (!relative.startsWith("..") && !path.isAbsolute(relative));
-}
-
-function pluginMcpConfigs(
-  plugins: RuntimePluginRef[],
-): Array<{ pluginName: string; configPath: string }> {
-  const seen = new Set<string>();
-  return plugins.flatMap((plugin) => {
-    const configPath =
-      plugin.mcpConfigPath ??
-      (plugin.path && !plugin.path.endsWith(".app") ? path.join(plugin.path, ".mcp.json") : null);
-    if (!configPath || !existsSync(configPath)) return [];
-    const resolved = path.resolve(configPath);
-    if (!shouldLoadMcpConfig(plugin, resolved)) return [];
-    if (seen.has(resolved)) return [];
-    seen.add(resolved);
-    return [
-      { pluginName: plugin.name || path.basename(path.dirname(resolved)), configPath: resolved },
-    ];
-  });
-}
-
-function deriveFrontendBase(): string {
-  const port = process.env.PORT || "3000";
-  return `http://127.0.0.1:${port}`;
 }
 
 async function fetchModelsFromBackend(settings: ApiSettings): Promise<AgentModel[]> {
